Started work on new simpler (less special constructs for special expressions) and more efficient grammar.
Instead of writing:

if(condition) {
    trueCaseCode
} else {
    falseCaseCode
}

The following can be written:

if {condition}, {
    trueCaseCode
} Else {
    falseCaseCode
}

which is interpreted as an ifElse message with three arguments to be
sent to self. In the compiler, recognition of such special message can
be inlined, i.e. a couple of hardcoded macros can handled this such that
efficiency is the same runtime-wise.

Signed-off-by: Jakob Brandsborg Ehmsen <jakobehmsen.github@gmail.com>