Cancelled duo-dict type approach for "dict-from-context". Instead, a single-dict approach is decided as follows:
The grammar for dicts persists - with a single extension: an optional
explicit prototype expression written as:

#thePrototypeExpression[...]

where [...] is derived from the resulting prototype from
thePrototypeExpression.

This implies that #[] is derived from an implicit prototype. So far,
this has been Any. It will now be changed to this. 

I.e.

#[...]

will correspond to

#this[...]

Signed-off-by: Jakob Brandsborg Ehmsen <jakobehmsen.github@gmail.com>