Added grammar for active process spawning:
#anEnvironment(...)

where environment is the process used as receiver when sending implicit
message to self.

environment is optional; if not expressed, then the implicit receiver
will be used.

Thus

#(...)

is also valid.

The objective is that: Active processes are to be for passive processes
as (more or less) coroutines are to procedures/functions.

Signed-off-by: Jakob Brandsborg Ehmsen <jakobehmsen.github@gmail.com>